Store ratio as number of elements and number of elements classified as spam.  Added method ratio() to get the calculated ratio.
Added messages when stopping processing of subtrees (~ terminating
recursion) to make it easier to see how the tree generation is
progressing.

Separate a validation set from the training set to allow checking the error
ratio of the classification against the set.  Randomly permutate rows
in the training set before training the classifier to choose the
validation set at random.

Added column names to nodes to make the tree easier to understand.